Moved Task system to lib & QoL changes
Moved most of the task system's functionality to the library for easy portability.
Moved a few properties to the LIBRARY_SETTINGS so they can easily be used in the library.
Many QoL changes for both user and developer.

From now on my aim is to start cleaning up the code & improving it where possible as some parts are outdated with unneeded or unoptimized code. If you see something that could be improved, pelase lmk.